backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket support
websockets==12.0

# AI/ML Services
groq==0.37.1
python-dotenv==1.0.0

# Utilities
requests==2.31.0
